modules/LocationConsumer/requirements.txt
psycopg2-binary==2.8.5
SQLAlchemy==1.3.19
kafka-python==2.0.2
requests==2.25.0